requirements.txt

python-telegram-bot[ext]>=20.3    # Бот для Telegram, поддержка расширенных функций
pydantic-settings==2.0.3          # Управление настройками через Pydantic
python-dotenv==1.0.0              # Загрузка переменных окружения из .env файлов
peewee==3.17.0                    # Лёгкая ORM для работы с базами данных
httpx==0.24.1                     # Асинхронный HTTP-клиент
loguru==0.7.0                     # Удобная библиотека для логирования
pydantic==2.11.1                  # Валидация данных и управление типами
pytest==8.1.1                     # Фреймворк для тестирования
pytest-asyncio==0.23.5            # Асинхронное тестирование с pytest
python-dateutil==2.8.2            # Работа с датами и временем
black==24.0.0                     # Форматтер кода
pylint==3.0.3                     # Анализатор кода
pytest-cov==4.1.0                 # Покрытие кода тестами для pytest
pytz==2023.3                      # Работа с часовыми поясами
apscheduler>=3.10.0               # Планировщик задач
cachetools==5.3.1                 # Кеширование